[MLDB-1443] explain function now takes a float as label. fixed glz explain segfault for regression

+++ ml/jml/boosted_stumps.h
@@ -255,7 +255,7 @@ public:
     merge(const Classifier_Impl & other, float weight = 1.0) const;
 
     virtual Explanation explain(const Feature_Set & feature_set,
-                                int label,
+                                float label,
                                 double weight = 1.0,
                                 PredictionContext * context = 0) const;
