+++ ml/jml/glz_classifier.h
@@ -181,7 +181,7 @@ protected:
 
 public:
     virtual Explanation explain(const Feature_Set & feature_set,
-                                int label,
+                                float label,
                                 double weight = 1.0,
                                 PredictionContext * context = 0) const;
